sysmon/requirements.txt
psutil
pandas
nvgpu
matplotlib
hydra-core

